Clarification for noobs (a.k.a. people like me)
I managed to get it to run. I was missing both the auth_secret (putting it in the environment variables), and the openrouter_api_key.

It seems the openrouter_api_key is absolutely necessary (without it just doesn´t answer (sends error)). This is kinda misleading because in the README it says:

This repo is compatible with OpenRouter and OpenAI. To use OpenRouter, you need to set the OPENROUTER_API_KEY environment variable.

This makes it sound like you need one or the other (which is what makes sense to me, as I understand it they´re 2 different LLM providers). I was putting only the open_ai_key at the beginning and it was failing, added an open_router and now it works